Сделал четвертое задание. Добавил разделение на подпроекты.
