+++ src/main/java/Main.java
@@ -1,36 +1,19 @@
+import argtoar.ArgToAr;
+import getmin.GetMin;
+
 public class Main {
-  int foo() { 
-    return 10;
-  }
 
-  static int getMin(int[] array) {
-    if (array.length == 0) {
-      return Integer.MAX_VALUE;
-    } else {
-      int value = Integer.MAX_VALUE;
-      for (int i = 0; i < array.length; i++) {
-        if (array[i] < value) {
-          value = array[i];
-        }
-      }
-      return value;
-    }
-  }
-
-  static String printMinOrError(String[] args) {
-    int[] curArray = new int[args.length];
-    for (int i = 0; i < args.length; i++) {
-      try {
-        curArray[i] = Integer.parseInt(args[i]);
-      } catch (NumberFormatException exc) {
-        return new String(args[i] + " is not a number");
-      }
-    }
-    return Integer.toString(getMin(curArray));
-  }
+  /**
+  * why do u need these docs?.
+  **/
 
   public static void main(String[] args) {
-    printMinOrError(args);    
-  }
-
+    int[] curArray = ArgToAr.get(args);
+    if (curArray == null) {
+      System.out.println("Wrong data");
+      return;
+    }
+    int answer = GetMin.getMin(curArray);
+    System.out.println(Integer.toString(answer));
+  }       
 }
